Moved one more exporter test from the view to exporter test file
