+++ evap/results/tests/test_exporters.py
@@ -2,8 +2,9 @@ import xlrd
 from io import BytesIO
 from model_mommy import mommy
 from django.test import TestCase
+from django.utils import translation
 
-from evap.evaluation.models import Semester, Course, Contribution, UserProfile, Question, Questionnaire, RatingAnswerCounter
+from evap.evaluation.models import Semester, Course, Contribution, UserProfile, Question, Questionnaire, RatingAnswerCounter, CourseType
 from evap.results.exporters import ExcelExporter
 
 
@@ -45,3 +46,36 @@ class TestExporters(TestCase):
         self.assertEqual(workbook.sheets()[0].row_values(3)[0], heading_question.text)
         self.assertEqual(workbook.sheets()[0].row_values(4)[0], likert_question.text)
         self.assertEqual(workbook.sheets()[0].row_values(5)[0], "")
+
+    def test_view_excel_file_sorted(self):
+        semester = mommy.make(Semester)
+        course_type = mommy.make(CourseType)
+        course1 = mommy.make(Course, state='published', type=course_type,
+                             name_de='A - Course1', name_en='B - Course1', semester=semester)
+
+        course2 = mommy.make(Course, state='published', type=course_type,
+                             name_de='B - Course2', name_en='A - Course2', semester=semester)
+
+        mommy.make(Contribution, course=course1, responsible=True, can_edit=True, comment_visibility=Contribution.ALL_COMMENTS)
+        mommy.make(Contribution, course=course2, responsible=True, can_edit=True, comment_visibility=Contribution.ALL_COMMENTS)
+
+        content_de = BytesIO()
+        with translation.override("de"):
+            ExcelExporter(semester).export(content_de, [[course_type.id]], True, True)
+
+        content_en = BytesIO()
+        with translation.override("en"):
+            ExcelExporter(semester).export(content_en, [[course_type.id]], True, True)
+
+        content_de.seek(0)
+        content_en.seek(0)
+
+        # Load responses as Excel files and check for correct sorting
+        workbook = xlrd.open_workbook(file_contents=content_de.read())
+        self.assertEqual(workbook.sheets()[0].row_values(0)[1], "A - Course1")
+        self.assertEqual(workbook.sheets()[0].row_values(0)[3], "B - Course2")
+
+        workbook = xlrd.open_workbook(file_contents=content_en.read())
+        self.assertEqual(workbook.sheets()[0].row_values(0)[1], "A - Course2")
+        self.assertEqual(workbook.sheets()[0].row_values(0)[3], "B - Course1")
+
